+++ contracts/interfaces/IStrategyManager.sol
@@ -12,8 +12,8 @@ interface IStrategyManager {
     error StrategyHasNoSelector();
 
     // Events
-    event NewStrategy(uint16 strategyId, address implementation);
-    event StrategyUpdated(uint16 strategyId, bool isActive, uint16 standardProtocolFee, uint16 minTotalFee);
+    event NewStrategy(uint256 strategyId, address implementation);
+    event StrategyUpdated(uint256 strategyId, bool isActive, uint16 standardProtocolFee, uint16 minTotalFee);
 
     // Custom structs
     struct Strategy {
